#26 - device classes added to subpackage __init__.py files
