+++ src/example_scripts/finite_multi_fifo_mode.py
@@ -3,7 +3,7 @@ integration test."""
 from typing import List, Optional
 
 from spectrumdevice import MockSpectrumDigitiserCard, SpectrumDigitiserCard
-from spectrumdevice.devices.measurement import Measurement
+from spectrumdevice.measurement import Measurement
 from spectrumdevice.settings import (
     AcquisitionMode,
     CardType,
