+++ src/example_scripts/standard_single_mode.py
@@ -5,7 +5,7 @@ from typing import Optional
 
 from spectrumdevice import MockSpectrumDigitiserCard, SpectrumDigitiserCard
 
-from spectrumdevice.devices.measurement import Measurement
+from spectrumdevice.measurement import Measurement
 from spectrumdevice.settings import (
     AcquisitionMode,
     CardType,
