+++ src/spectrumdevice/devices/digitiser/__init__.py
@@ -1,0 +1,15 @@
+from spectrumdevice.devices.digitiser.digitiser_card import SpectrumDigitiserCard
+from spectrumdevice.devices.digitiser.digitiser_channel import SpectrumDigitiserChannel
+from spectrumdevice.devices.digitiser.digitiser_interface import (
+    SpectrumDigitiserChannelInterface,
+    SpectrumDigitiserInterface,
+)
+from spectrumdevice.devices.digitiser.digitiser_star_hub import SpectrumDigitiserStarHub
+
+__all__ = [
+    "SpectrumDigitiserChannelInterface",
+    "SpectrumDigitiserChannel",
+    "SpectrumDigitiserInterface",
+    "SpectrumDigitiserCard",
+    "SpectrumDigitiserStarHub",
+]
