+++ src/spectrumdevice/devices/mocks/__init__.py
@@ -5,25 +5,19 @@
 # Licensed under the MIT. You may obtain a copy at https://opensource.org/licenses/MIT.
 
 import logging
-from abc import ABC, abstractmethod
-from threading import Event, Lock, Thread
-from time import monotonic, sleep
+from time import sleep
 from typing import List, Optional, Sequence
 
-from numpy import ndarray, zeros
-from numpy.random import uniform
+from numpy import zeros
 
 from spectrum_gmbh.regs import (
-    SPC_CARDMODE,
     SPC_MIINST_CHPERMODULE,
-    SPC_MIINST_MAXADCVALUE,
     SPC_MIINST_MODULES,
     SPC_PCITYP,
 )
-from spectrumdevice.devices.digitiser.abstract_spectrum_digitiser import AbstractSpectrumDigitiser
-from spectrumdevice.devices.digitiser.digitiser_card import SpectrumDigitiserCard
-from spectrumdevice.devices.digitiser.digitiser_star_hub import SpectrumDigitiserStarHub
-from spectrumdevice.devices.mocks.mock_abstract_spectrum_device import MockAbstractSpectrumDevice
+from spectrumdevice.devices.digitiser import SpectrumDigitiserCard
+from spectrumdevice.devices.digitiser import SpectrumDigitiserStarHub
+from spectrumdevice.devices.mocks.mock_abstract_devices import MockAbstractSpectrumDigitiser
 from spectrumdevice.devices.mocks.timestamps import MockTimestamper
 from spectrumdevice.exceptions import (
     SpectrumNoTransferBufferDefined,
@@ -34,55 +28,6 @@ from spectrumdevice.settings.device_modes import AcquisitionMode
 from spectrumdevice.settings.transfer_buffer import CardToPCDataTransferBuffer
 
 logger = logging.getLogger(__name__)
-
-
-class MockAbstractSpectrumDigitiser(MockAbstractSpectrumDevice, AbstractSpectrumDigitiser, ABC):
-    """Overrides methods of `AbstractSpectrumDigitiser` that communicate with hardware with mocked implementations, allowing
-    software to be tested without Spectrum hardware connected or drivers installed, e.g. during CI. Instances of this
-    class cannot be constructed directly - instantiate `MockAbstractSpectrumDigitiser` and `MockSpectrumStarHub` objects instead,
-    which inherit from this class."""
-
-    def __init__(self, source_frame_rate_hz: float = 10.0) -> None:
-        """
-        Args:
-            source_frame_rate_hz (float): Frame rate at which a mock waveform source will generate waveforms.
-        """
-        MockAbstractSpectrumDevice.__init__(self)
-        self._source_frame_rate_hz = source_frame_rate_hz
-        self._param_dict[SPC_CARDMODE] = AcquisitionMode.SPC_REC_STD_SINGLE.value
-        self._param_dict[SPC_MIINST_MAXADCVALUE] = 128
-
-        self._buffer_lock = Lock()
-        self._acquisition_stop_event = Event()
-        self._acquisition_thread: Optional[Thread] = None
-        self._timestamp_thread: Optional[Thread] = None
-        self._enabled_channels = [0]
-        self._on_device_buffer: ndarray = zeros(1000)
-        self._previous_data = self._on_device_buffer.copy()
-
-    def start(self) -> None:
-        """Starts a mock waveform source in a separate thread. The source generates noise samples according to the
-        number of currently enabled channels and the acquisition length, and places them in the virtual on abstract_device buffer
-        (the _on_device_buffer attribute).
-        """
-        waveform_source = mock_waveform_source_factory(self.acquisition_mode)
-        amplitude = self.read_spectrum_device_register(SPC_MIINST_MAXADCVALUE)
-        self._acquisition_stop_event.clear()
-        self._acquisition_thread = Thread(
-            target=waveform_source,
-            args=(
-                self._acquisition_stop_event,
-                self._source_frame_rate_hz,
-                amplitude,
-                self._on_device_buffer,
-                self._buffer_lock,
-            ),
-        )
-        self._acquisition_thread.start()
-
-    def stop(self) -> None:
-        """Stops the mock waveform source and timestamp threads."""
-        self._acquisition_stop_event.set()
 
 
 class MockSpectrumDigitiserCard(SpectrumDigitiserCard, MockAbstractSpectrumDigitiser):
@@ -271,70 +216,3 @@ class MockSpectrumDigitiserStarHub(SpectrumDigitiserStarHub, MockAbstractSpectru
         """
         for card in self._child_cards:
             card.stop()
-
-
-class MockWaveformSource(ABC):
-    """Interface for a mock noise waveform source. Implementations are intended to be called in their own thread.
-    When called, `MockWaveformSource` implementations will fill a provided buffer with noise samples."""
-
-    @abstractmethod
-    def __call__(
-        self, stop_flag: Event, frame_rate: float, amplitude: float, on_device_buffer: ndarray, buffer_lock: Lock
-    ) -> None:
-        raise NotImplementedError()
-
-
-class SingleModeMockWaveformSource(MockWaveformSource):
-    def __call__(
-        self, stop_flag: Event, frame_rate: float, amplitude: float, on_device_buffer: ndarray, buffer_lock: Lock
-    ) -> None:
-        """When called, this MockWaveformSource simulates SPC_REC_STD_SINGLE Mode, placing a single frames worth of
-        samples into a provided mock on_device_buffer.
-
-        Args:
-            stop_flag (Event): A threading event that will be used in the calling thread to stop the acquisition.
-            frame_rate (float): The samples will be generated 1 / frame_rate seconds after __call__ is called.
-            amplitude (float): Waveforms will contain random values in the range -amplitude to +amplitude
-            on_device_buffer (ndarray): The numpy array into which the noise samples will be written.
-            buffer_lock (Lock): A threading lock created in the calling thread that will be used to ensure access to
-                the on_device_buffer array is thread safe.
-
-        """
-        start_time = monotonic()
-        while not stop_flag.is_set() and ((monotonic() - start_time) < (1 / frame_rate)):
-            sleep(0.001)
-        if not stop_flag.is_set():
-            with buffer_lock:
-                on_device_buffer[:] = uniform(low=-1 * amplitude, high=amplitude, size=len(on_device_buffer))
-
-
-class MultiFIFOModeMockWaveformSource(MockWaveformSource):
-    def __call__(
-        self, stop_flag: Event, frame_rate: float, amplitude: float, on_device_buffer: ndarray, buffer_lock: Lock
-    ) -> None:
-        """When called, this `MockWaveformSource` simulates SPC_REC_FIFO_MULTI Mode, continuously replacing the contents
-        of on_device_buffer with new frames of noise samples.
-
-        Args:
-            stop_flag (Event): A threading event that will be used in the calling thread to stop the acquisition.
-            frame_rate (float): The contents of the on_device_buffer will be updated at this rate (Hz).
-            amplitude (float): Waveforms will contain random values from a uniform distribution in the range -amplitude
-            to +amplitude
-            on_device_buffer (ndarray): The numpy array into which the noise samples will be written.
-            buffer_lock (Lock): A threading lock created in the calling thread that will be used to ensure access to
-                the on_device_buffer array is thread safe.
-
-        """
-        while not stop_flag.is_set():
-            with buffer_lock:
-                on_device_buffer[:] = uniform(low=-1 * amplitude, high=amplitude, size=len(on_device_buffer))
-                sleep(1 / frame_rate)
-
-
-def mock_waveform_source_factory(acquisition_mode: AcquisitionMode) -> MockWaveformSource:
-    if acquisition_mode == AcquisitionMode.SPC_REC_FIFO_MULTI:
-        return MultiFIFOModeMockWaveformSource()
-    elif AcquisitionMode.SPC_REC_STD_SINGLE:
-        return SingleModeMockWaveformSource()
-    else:
-        raise NotImplementedError(f"Mock waveform source not yet implemented for {acquisition_mode} acquisition mode.")
